fix: address cubic review findings on Granola templates
- folders sync: only (re)start the deletion tracking window on a
  fresh sync; resuming from a saved cursor no longer wipes out
  folders from pages already saved before the interruption
- notes sync and get-note action: map through space_membership and
  folder_membership[].space_id, which the API returns but the
  schemas silently dropped
- list-folders, list-notes, get-transcript actions: require page_size
  to be an integer (and within range for list-folders) before the
  request is sent, verified against the live Granola API which
  rejects fractional/out-of-range values with 400s

Verified space_membership/space_id and the page_size validation
behavior against the live Granola API via a real Nango connection.

+++ integrations/granola/actions/get-transcript.ts
@@ -5,7 +5,7 @@ const InputSchema = z
     .object({
         note_id: z.string().describe('Unique identifier of the note whose transcript to retrieve. Example: "not_IBuThyr2eJgage"'),
         cursor: z.string().optional().describe('Opaque pagination cursor from a previous response. Omit for the first page.'),
-        page_size: z.number().min(1).max(100).optional().describe('Number of transcript segments per page. Maximum 100, default 50.')
+        page_size: z.number().int().min(1).max(100).optional().describe('Number of transcript segments per page. Maximum 100, default 50.')
     })
     .describe('Input for retrieving a paginated meeting transcript.');
